// src/components/button/types.ts
export const paddingSmall = '8px 35px'
export const paddingMedium = '14px 74px'
export const paddingLarge = '18px 74px'

export type TextType = 'white' | '#1B4963' | 'linear-gradient(95.85deg, #09152F 0%, #1B4963 100%)'
export type MainColor = '#1B4963' | 'linear-gradient(95.85deg, #09152F 0%, #1B4963 100%)'

export enum TypeOfButton {
    filled = 'filled',
    circled = 'circled',
    simple = 'simple'
}

export enum SizeOfButton {
    small = 'small',
    medium = 'medium',
    large = 'large'
}

export enum BackgroundTypeOfButton {
    simple = 'simple',
    gradient = 'gradient',
}